fix error with eureka and FeignClient

+++ passengers/src/main/java/be/vinci/ipl/projet/passengers/data/PassengersRepository.java
@@ -25,5 +25,5 @@ public interface PassengersRepository extends CrudRepository<Passenger, Integer>
   void deleteByTripId(int tripId);
 
   @Transactional
-  void deleteByUserId(long userId);
+  void deleteByUserId(int userId);
 }
